+++ tests/src/Rector/PHPUnit/PhpUnitTestAnnotationToAttributeRector/PhpUnitTestAnnotationToAttributeRectorTest.php
@@ -1,0 +1,31 @@
+<?php
+
+declare(strict_types=1);
+
+namespace DrupalRector\Tests\Rector\PHPUnit\PhpUnitTestAnnotationToAttributeRector;
+
+use DrupalRector\Tests\AbstractDrupalRectorTestCase;
+use Iterator;
+
+#[\PHPUnit\Framework\Attributes\CoversFunction('refactor')]
+class PhpUnitTestAnnotationToAttributeRectorTest extends AbstractDrupalRectorTestCase
+{
+    #[\PHPUnit\Framework\Attributes\DataProvider('provideData')]
+    public function test(string $filePath): void
+    {
+        $this->doTestFile($filePath);
+    }
+
+    /**
+     * @return Iterator<<string>>
+     */
+    public static function provideData(): \Iterator
+    {
+        return self::yieldFilesFromDirectory(__DIR__.'/fixture');
+    }
+
+    public function provideConfigFilePath(): string
+    {
+        return __DIR__.'/config/configured_rule.php';
+    }
+}